feat: add project deletion and update project repository
This commit introduces the ability to delete projects by their ID in `ProjectRepositoryImpl`. It also adds a `deleteProject` function to the `ProjectRepository` interface. In `ProjectRepositoryImpl`, a new function `deleteProject` has been implemented which remove a project from the projects list and return success or failure based on whether the project exists or not. Also `editProject` and `getProjectById` methods has been modified.